debug a lot of multer and sharp uploader system

+++ route/uploader_middleware.js
@@ -3,8 +3,7 @@ const sharp = require('sharp');
 const path = require('path');
 const fs = require('fs');
 
-
-const globalMulterUploader = ({ path, size, filter }) => {
+const globalMulterUploader = ({ name, path, size, filter }) => {
   return (req, res, next) => {
     const upload = multer({
       storage: multer.diskStorage({
@@ -13,31 +12,47 @@ const globalMulterUploader = ({ path, size, filter }) => {
       }),
       limits: { fileSize: size },
       fileFilter: (req, file, cb) => {
-        if (file.mimetype.startsWith(filter)) {
+        const filterArray = filter.some((item) => file.mimetype.includes(item));
+        if (filterArray) {
           cb(null, true);
         } else {
-          cb(new Error(`File type only ${filter} and up to ${(size / 1024)}kb`));
+          cb(new Error(`File type only ${filter.join(', ')} and up to ${(size / 1024)}kb`));
         }
       }
-    }).any('files');
+    });
 
-    upload(req, res, next, (err) => {
-      if (err) {
-        next(err.message)
+    const uploadHandler = Array.isArray(name) && name[0] === 'single' 
+      ? upload.single(name[1]) 
+      : upload.any('files');
+
+    uploadHandler(req, res, (err) => {
+      if (err && err instanceof multer.MulterError && err.code === 'LIMIT_UNEXPECTED_FILE') {
+        next(); // No file uploaded, continue without error
+      } else if (err) {
+       return next('Multer:'+ err.message)
+        // return res.status(400).send({ msg: 'Multer:'+ err.message, alert: "alert-danger" });
+      } else {
+        next();
       }
-      next();
     });
   };
 };
 
 
-const globalSharpReducer = ({ quality=50 }) => {
+const globalSharpReducer = ({ quality = 50}) => {
   return async (req, res, next) => {
+    
+    const files = req.files || (req.file ? [req.file] : []);
+    
+    if (!files.length) {
+      return next(); // No files to process, continue without error
+    }
+
     try {
-      const resizePromises = req.files.map(async (file) => {
+      const resizePromises = files.map(async (file) => {
         const outputPath = path.resolve(file.destination, 'resized', file.filename);
         await sharp(file.path)
-        //   .resize({ width, height, fit: sharp.fit.cover, withoutEnlargement: true })
+          // .resize({ width, height, fit: sharp.fit.cover, withoutEnlargement: true })
           .jpeg({ quality })
           .toFile(outputPath);
         fs.unlinkSync(file.path);
@@ -46,10 +61,12 @@ const globalSharpReducer = ({ quality=50 }) => {
       await Promise.all(resizePromises);
       next();
     } catch (error) {
-      next(error);
+      next('sharp: '+error);
     }
   };
 };
+
+
 
 module.exports={
     globalMulterUploader, globalSharpReducer
